end page room-detail
