+++ src/js/index.js
@@ -85,7 +85,21 @@ $(function() {
 	const totalPrice = $('.js-total-price');
 	let total = 0;
 	let dayOrDays = 'сутки';
+	const btnDatepicker = $('.js-btn-datepicker');
+	const inputDatepicker = $('.js-input-datepicker');
+	const tglDrop = $('.js-tgl-dropdown');
+	const dropdown = $('.js-dropdown-form');
+	const dropdownOpen = 'formfield__dropdown_open';
 
+	// Declaration
+	const declOfNum = (number, titles) => {
+		return (
+			titles[(number % 100 > 4 && number % 100 < 20) ?
+				2 : cases[(number % 10 < 5) ? number % 10 : 5]]
+		);
+	}
+
+	// Дополнительные настройки + видмость кнопок
 	calendar.datepicker({
 		showOtherMonths: true,
 		prevText: '',
@@ -104,8 +118,10 @@ $(function() {
 		}
 	});
 
+	// Обнуляю при загрузке
 	calendar.datepicker('setDate', [null, null]);
 
+	// Л=Добавляю свою кнопки
 	calendar.append(`
 		<div class="datepicker-my-btns">
 			<button class="btn-purple btn-purple_visible js-reset" type="button">очистить</button>
@@ -113,6 +129,7 @@ $(function() {
 		</div>
 	`);
 
+	// Очистка календаря
 	$(document).on('click', '.js-reset', () => {
 		const extensionRange = calendar.datepicker('widget').data('datepickerExtensionRange');
 		extensionRange.startDateText = null;
@@ -129,6 +146,7 @@ $(function() {
 		totalPrice.text(0);
 	});
 
+	// Заполнение дат, подсчёт стоимости
 	$(document).on('click', '.js-apply', () => {
 		const extensionRange = calendar.datepicker('widget').data('datepickerExtensionRange');
 		const start = new Date(extensionRange.startDate).getTime();
@@ -145,12 +163,7 @@ $(function() {
 		calendar.hide();
 	});
 
-	const btnDatepicker = $('.js-btn-datepicker');
-	const inputDatepicker = $('.js-input-datepicker');
-	const tglDrop = $('.js-tgl-dropdown');
-	const dropdown = $('.js-dropdown-form');
-	const dropdownOpen = 'formfield__dropdown_open';
-
+	// Показ календаря
 	btnDatepicker.on('click', function() {
 		$(this).siblings('.js-input-datepicker').focus();
 	});
@@ -174,13 +187,6 @@ $(function() {
 			}
 		}
 	});
-
-	const declOfNum = (number, titles) => {
-		return (
-			titles[(number % 100 > 4 && number % 100 < 20) ?
-				2 : cases[(number % 10 < 5) ? number % 10 : 5]]
-		);
-	}
 
 	// Открытие / закрытие выпадающего списка
 	tglDrop.on('click', function() {
@@ -217,6 +223,7 @@ $(function() {
 		}
 	});
 
+	// Закрытие выпадающего списка по кнопке - Применить
 	$('.js-dropdown-apply').on('click', () => {
 		if (countPeople !== 0) {
 			guestsSelector.text(`${countPeople} ${declOfNum(countPeople, guestsDeclaration)}`);
@@ -226,6 +233,7 @@ $(function() {
 		dropdown.removeClass(dropdownOpen);
 	});
 
+	// Сброс значений выпадающего списка
 	dropdownReset.on('click', function() {
 		$('.js-current-total').val(0);
 		countPeople = 0;
@@ -233,126 +241,13 @@ $(function() {
 		$(this).removeClass('btn-purple_visible')
 	});
 
-	let SwitchPopup = window.SwitchPopup || {};
-
-	SwitchPopup = (function () {
-		function SwitchPopup(element, settings) {
-			const thiz = this;
-
-			thiz.$popup = $(element);
-			thiz.state = 'close';
-			thiz.pageScrollClass = 'html';
-			thiz.btnClass = '';
-			thiz.duration = 300;
-			thiz.overflow = true;
-			thiz.displayClass = 'popup_display';
-			thiz.visibleClass = 'popup_visible';
-			thiz.$scrollWidth = $(':root').css('--scroll-width');
-
-			thiz.options = $.extend(thiz, settings);
-
-			thiz.init();
-		}
-		return SwitchPopup;
-	})();
-
-	SwitchPopup.prototype.open = function () {
-		const thiz = this;
-		thiz.state = 'opening';
-		thiz.$popup.trigger('beforeOpen', [thiz.$popup]);
-		thiz.$popup.trigger('beforeChange', [thiz.$popup, 'open']);
-		thiz.$popup.addClass(thiz.displayClass);
-		setTimeout(function () {
-			thiz.$popup.addClass(thiz.visibleClass);
-		}, 1);
-		setTimeout(function () {
-			thiz.state = 'open';
-			thiz.$popup.trigger('afterOpen', [thiz.$popup]);
-			thiz.$popup.trigger('afterChange', [thiz.$popup, 'open']);
-		}, thiz.duration);
-		if (thiz.overflow) {
-			$(thiz.pageScrollClass).css({
-				position: 'fixed',
-				top: 0,
-				paddingRight: thiz.$scrollWidth,
-				overflow: 'hidden'
-			});
-		}
-	};
-
-	SwitchPopup.prototype.close = function () {
-		const thiz = this;
-		thiz.state = 'closing';
-		thiz.$popup.trigger('beforeClose', [thiz.$popup]);
-		thiz.$popup.trigger('beforeChange', [thiz.$popup, 'close']);
-		thiz.$popup.removeClass(thiz.visibleClass);
-		setTimeout(function () {
-			thiz.$popup.removeClass(thiz.displayClass);
-			thiz.state = 'close';
-			thiz.$popup.trigger('afterClose', [thiz.$popup]);
-			thiz.$popup.trigger('afterChange', [thiz.$popup, 'close']);
-			if (thiz.overflow) {
-				$(thiz.pageScrollClass).css({
-					position: 'relative',
-					top: 'auto',
-					paddingRight: 0,
-					overflow: 'auto'
-				});
-			}
-		}, thiz.duration);
-	};
-
-	SwitchPopup.prototype.toggle = function () {
-		const thiz = this;
-		if (thiz.$popup.hasClass(thiz.displayClass)) {
-			thiz.close();
-		} else {
-			thiz.open();
-		}
-	};
-
-	SwitchPopup.prototype.getState = function () {
-		const thiz = this;
-		return thiz.state;
-	};
-
-	SwitchPopup.prototype.init = function () {
-		const thiz = this;
-		if (thiz.$popup.length) {
-			if (!thiz.btnClass.length) return;
-			$(document).on('click', '.' + thiz.btnClass, function () {
-				thiz.toggle();
-			});
-		}
-		thiz.$popup.trigger('init', [thiz.$popup]);
-	};
-
-	$.fn.switchPopup = function () {
-		const thiz = this;
-		const opt = arguments[0];
-		const args = Array.prototype.slice.call(arguments, 1);
-		const l = thiz.length;
-		let i;
-		let ret;
-
-		for (i = 0; i < l; i++) {
-			if (typeof opt === 'object' || typeof opt === 'undefined') {
-				thiz[i].switchPopup = new SwitchPopup(thiz[i], opt);
-			} else {
-				ret = thiz[i].switchPopup[opt].apply(thiz[i].switchPopup, args);
-			}
-			if (typeof ret !== 'undefined') return ret;
-		}
-		return thiz;
-	};
-
 	// Функция для инициализации попапов
 	function initPopup(popup, btn) {
 		popup.switchPopup({
 			pageScrollClass: '.root',
 			btnClass: btn,
 			duration: 300,
-			overflow: true,
+			overflow: true
 		});
 	}
